+++ packages/website/src/components/DriveFileName.tsx
@@ -1,7 +1,9 @@
 import { Stack } from 'office-ui-fabric-react';
 import React from 'react';
+import { Link } from 'react-router-dom';
 import { DriveFile, mdLink, MimeTypes } from '../utils';
-import { ShortcutIcon } from '.';
+import { DriveIcon } from './DriveIcon';
+import { ShortcutIcon } from './ShortcutIcon';
 
 function DriveFileName_({ file }: { file: DriveFile }) {
   const link = mdLink.parse(file.name);
@@ -19,3 +21,32 @@ function DriveFileName_({ file }: { file: DriveFile }) {
 }
 
 export const DriveFileName = React.memo(DriveFileName_);
+
+export interface IFileInList {
+  file: DriveFile;
+  openInNewWindow: boolean;
+  children: any;
+}
+
+export function FileWithIcon({ file }: { file: DriveFile }) {
+  return (
+    <Stack verticalAlign="center" horizontal tokens={{ childrenGap: 8 }}>
+      <DriveIcon file={file} />
+      <DriveFileName file={file} />
+    </Stack>
+  );
+}
+
+export function FileLink({ file, openInNewWindow, children }: IFileInList) {
+  const link = mdLink.parse(file.name);
+  const target = openInNewWindow ? '_blank' : undefined;
+  return link ? (
+    <a href={link.url} target="_blank" rel="noreferrer">
+      {children}
+    </a>
+  ) : (
+    <Link to={`/view/${file.id}`} target={target}>
+      {children}
+    </Link>
+  );
+}
